feat(webrtc): allow to customize offer (#9)
In certain cases (e.g. React Native) we'd need to make adjustments to
the offer that aren't currently happening.

We might do it internally in the SDK at the future, but for now let's
make it the consumers responsibility, and just give them the tools to do
so.

If we'll see additional similar needs - let's design a proper solution